+++ streaming.c
@@ -184,7 +184,7 @@ struct chunkID_set *cb_to_bmap(struct chunk_buffer *chbuf)
 }
 
 // a simple implementation that request everything that we miss ... up to max deliver
-struct chunkID_set *get_chunks_to_accept(struct nodeID *fromid, const struct chunkID_set *cset_off, int max_deliver, int trans_id){
+struct chunkID_set *get_chunks_to_accept(struct nodeID *fromid, const struct chunkID_set *cset_off, int max_deliver, uint16_t trans_id){
   struct chunkID_set *cset_acc, *my_bmap;
   int i, d, cset_off_size;
   //double lossrate;
@@ -435,7 +435,7 @@ double getChunkTimestamp(int *cid){
   return (double) c->timestamp;
 }
 
-void send_accepted_chunks(struct nodeID *toid, struct chunkID_set *cset_acc, int max_deliver, int trans_id){
+void send_accepted_chunks(struct nodeID *toid, struct chunkID_set *cset_acc, int max_deliver, uint16_t trans_id){
   int i, d, cset_acc_size, res;
   struct peer *to = nodeid_to_peer(toid, 0);
